# services/search/python_service/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AgenticAiETH Search Gateway",
    "version": "0.1.0"
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Read Health",
        "operationId": "read_health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/documents": {
      "post": {
        "summary": "Add Document",
        "operationId": "add_document_v1_documents_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchDocument"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/search": {
      "post": {
        "summary": "Search Documents",
        "operationId": "search_documents_v1_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/ai-search": {
      "post": {
        "summary": "Ai Search",
        "description": "AI-powered conversational search with source citations",
        "operationId": "ai_search_v1_ai_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AISearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AISearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/conversation/{conversation_id}": {
      "get": {
        "summary": "Get Conversation",
        "description": "Get conversation history",
        "operationId": "get_conversation_v1_conversation__conversation_id__get",
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Conversation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Conversation"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/sources/{doc_id}": {
      "get": {
        "summary": "Get Source Document",
        "description": "Get full source document by ID",
        "operationId": "get_source_document_v1_sources__doc_id__get",
        "parameters": [
          {
            "name": "doc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Doc Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Source Document V1 Sources  Doc Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AISearchRequest": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query"
          },
          "conversation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversation Id"
          },
          "max_sources": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Max Sources",
            "default": 5
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "AISearchRequest"
      },
      "AISearchResponse": {
        "properties": {
          "response": {
            "type": "string",
            "title": "Response"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/SearchResult"
            },
            "type": "array",
            "title": "Sources"
          },
          "conversation_id": {
            "type": "string",
            "title": "Conversation Id"
          },
          "follow_up_suggestions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Follow Up Suggestions"
          }
        },
        "type": "object",
        "required": [
          "response",
          "sources",
          "conversation_id"
        ],
        "title": "AISearchResponse"
      },
      "Conversation": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/ConversationMessage"
            },
            "type": "array",
            "title": "Messages"
          },
          "created_at": {
            "type": "number",
            "title": "Created At"
          },
          "last_accessed": {
            "type": "number",
            "title": "Last Accessed"
          }
        },
        "type": "object",
        "required": [
          "id",
          "messages",
          "created_at",
          "last_accessed"
        ],
        "title": "Conversation"
      },
      "ConversationMessage": {
        "properties": {
          "role": {
            "type": "string",
            "title": "Role"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "timestamp": {
            "type": "number",
            "title": "Timestamp"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/SearchResult"
            },
            "type": "array",
            "title": "Sources"
          }
        },
        "type": "object",
        "required": [
          "role",
          "content",
          "timestamp"
        ],
        "title": "ConversationMessage"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HealthResponse": {
        "properties": {
          "mode": {
            "type": "string",
            "title": "Mode"
          },
          "go_search": {
            "type": "boolean",
            "title": "Go Search"
          },
          "elasticsearch": {
            "type": "boolean",
            "title": "Elasticsearch"
          },
          "redis": {
            "type": "boolean",
            "title": "Redis"
          },
          "hana": {
            "type": "boolean",
            "title": "Hana"
          }
        },
        "type": "object",
        "required": [
          "mode",
          "go_search",
          "elasticsearch",
          "redis",
          "hana"
        ],
        "title": "HealthResponse"
      },
      "SearchDocument": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "id",
          "content"
        ],
        "title": "SearchDocument"
      },
      "SearchRequest": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query"
          },
          "top_k": {
            "type": "integer",
            "maximum": 200.0,
            "minimum": 1.0,
            "title": "Top K",
            "default": 10
          },
          "filters": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Filters"
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "SearchRequest"
      },
      "SearchResponse": {
        "properties": {
          "backend": {
            "type": "string",
            "title": "Backend"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/SearchResult"
            },
            "type": "array",
            "title": "Results"
          }
        },
        "type": "object",
        "required": [
          "backend",
          "results"
        ],
        "title": "SearchResponse"
      },
      "SearchResult": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "similarity": {
            "type": "number",
            "title": "Similarity"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "id",
          "content",
          "similarity"
        ],
        "title": "SearchResult"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}